chore(share): centralize shareId parsing util; guard share route logs; a11y/aria cleanup; robust header parsing; extract handleCreateShare; blob SDK var rename; indentation fix
