+++ src/components/ShareControls.tsx
@@ -6,6 +6,7 @@ import { fetchWithVercelBypass } from '@/utils/fetchWithVercelBypass';
 import { saveActivities } from '@/utils/activity-storage';
 import { importActivities } from '@/utils/activity-import-export';
 import type { PartialActivityImport } from '@/utils/activity-import-export';
+import { extractShareIdFromUrl } from '@/utils/shareUrl';
 
 interface Props {
   shareUrl: string;
@@ -14,30 +15,6 @@ interface Props {
 export default function ShareControls({ shareUrl }: Props) {
   const { addResponsiveToast } = useResponsiveToast();
 
-  // Extracted utility to parse share id from a share URL's last path segment
-  const extractShareIdFromUrl = (url: string): string | null => {
-    try {
-      if (!url || typeof url !== 'string') return null;
-      // Use URL parsing when possible; fallback to simple split for relative paths
-      const base = typeof window !== 'undefined' ? window.location.origin : 'http://localhost';
-      const u = new URL(url, base);
-      const pathname = u.pathname || '';
-      const segments = pathname.split('/').filter(Boolean);
-      const id = segments.length ? segments[segments.length - 1] : '';
-      return id && id.length > 0 ? id : null;
-    } catch {
-      try {
-        const normalized = url.trim();
-        const withoutHash = normalized.split('#')[0] || '';
-        const withoutQuery = withoutHash.split('?')[0] || '';
-        const parts = withoutQuery.split('/').filter(Boolean);
-        const id = parts.length ? parts[parts.length - 1] : '';
-        return id && id.length > 0 ? id : null;
-      } catch {
-        return null;
-      }
-    }
-  };
 
   const copy = async () => {
     try {
